gateway/requirements.txt
astroid==2.15.3
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dill==0.3.6
dnspython==2.3.0
Flask==2.2.3
Flask-PyMongo==2.3.0
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pika==1.3.1
platformdirs==3.2.0
pylint==2.17.2
pymongo==4.3.3
python-dotenv==1.0.0
requests==2.28.2
tomlkit==0.11.7
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
